ecochef_lite/chef.py
import streamlit as st
import google.generativeai as genai
import os

# Configure the Generative AI model
# You'll need to set your GOOGLE_API_KEY as an environment variable
# genai.configure(api_key=os.environ.get("GOOGLE_API_KEY"))

def generate_recipe_with_gemini(inventory_items, user_preferences=None):
    """
    Placeholder function to simulate generating a recipe with Gemini 1.5 Flash.
    In a real implementation, this would call the Google Generative AI SDK with
    an appropriate prompt.
    """
    st.info("Simulating recipe generation with Gemini 1.5 Flash...")

    # Construct a simple prompt based on inventory
    inventory_list = ", ".join([item[1] for item in inventory_items]) # item[1] is item_name
    
    prompt = f"Create a recipe using ONLY these ingredients: {inventory_list}. "
    if user_preferences:
        prompt += f"Prioritize {user_preferences['prioritize_item']}. Cuisine: {user_preferences['cuisine']}."
    prompt += "Output the recipe in a markdown-formatted card with sections for Ingredients, Instructions, and Notes."

    st.markdown(f"**Generated Prompt (for Gemini):**
```
{prompt}
```")

    # Dummy recipe for now
    dummy_recipe = """
**[Dummy Recipe] Delicious Pantry Stir-fry**

### Ingredients:
*   1 Avocado
*   500g Pasta
*   (Any other items you added)

### Instructions:
1.  Cook pasta according to package directions.
2.  While pasta is cooking, dice the avocado.
3.  Combine cooked pasta and avocado in a pan.
4.  Stir-fry for 2-3 minutes.
5.  Season to taste.

### Notes:
*   This recipe is a placeholder. In a real scenario, Gemini 1.5 Flash would generate a creative recipe based on your inventory.
*   Consider adding a protein or more vegetables for a balanced meal.
"""
    return dummy_recipe

def chef_module(inventory_items):
    st.header("👨‍🍳 Chef's Logic")

    if not inventory_items:
        st.warning("Your pantry is empty! Add some items to generate a recipe.")
        return

    st.write("Based on your current inventory, let's generate a recipe!")

    # User preferences for recipe generation
    st.subheader("Recipe Preferences (Optional)")
    col1, col2 = st.columns(2)
    with col1:
        prioritize_item = st.selectbox("Prioritize using an expiring item:", 
                                       options=["None"] + [item[1] for item in inventory_items if item[3] is not None and item[3] < 7]) # Assuming item[3] is days_until_expiry
    with col2:
        cuisine = st.text_input("Preferred Cuisine (e.g., Italian, Mexican, Asian)")

    user_preferences = {
        "prioritize_item": prioritize_item if prioritize_item != "None" else None,
        "cuisine": cuisine if cuisine else None
    }

    if st.button("Generate Recipe"):
        recipe = generate_recipe_with_gemini(inventory_items, user_preferences)
        st.subheader("Your Custom EcoChef Recipe:")
        st.markdown(recipe)

    st.write("---")
    st.write("Future features: Save recipe, adjust serving size, dietary filters.")
